move optimizer code into generator
