+++ packages/ts/jns42-generator/src/schema-transforms/alias.test.ts
