+++ packages/ts/jns42-generator/src/schema-transforms/flatten.test.ts
