+++ packages/ts/jns42-generator/src/schema-transforms/flip-all-of-one-of.test.ts
