+++ packages/ts/jns42-generator/src/schema-transforms/single-type.ts
