+++ packages/ts/jns42-generator/src/schema/arena.ts
