+++ packages/ts/jns42-generator/src/schema/index.ts
