+++ packages/ts/jns42-generator/src/schema/model.ts
