+++ packages/ts/jns42-generator/src/utils/arena.ts
