+++ packages/ts/jns42-generator/src/utils/calculate-hash.test.ts
