+++ packages/ts/jns42-generator/src/utils/empty.ts
