+++ packages/ts/jns42-generator/src/utils/index.ts
@@ -1,11 +1,21 @@
+export * from "./arena.js";
 export * from "./banner.js";
+export * from "./calculate-hash.js";
+export * from "./choose.js";
 export * from "./comments.js";
+export * from "./deep-equal.js";
+export * from "./deep-unique.js";
+export * from "./empty.js";
 export * from "./iterable-text-template.js";
 export * from "./iterable.js";
 export * from "./load.js";
+export * from "./math.js";
+export * from "./merge.js";
 export * from "./name.js";
 export * from "./namer.js";
+export * from "./normalize-object.js";
 export * from "./package.js";
+export * from "./properties.js";
 export * from "./read-node.js";
 export * from "./root.js";
 export * from "./schema.js";
