+++ packages/ts/jns42-generator/src/utils/math.test.ts
