+++ packages/ts/jns42-generator/src/utils/math.ts
