+++ packages/ts/jns42-generator/src/utils/normalize-object.test.ts
