+++ packages/ts/jns42-generator/src/utils/properties.test.ts
