+++ packages/ts/jns42-generator/src/utils/properties.ts
