--- packages/ts/jns42-optimizer/.npmrc
@@ -1,1 +1,0 @@
-package-lock=false
