--- packages/ts/jns42-optimizer/.vscode/launch.json
@@ -1,22 +1,0 @@
-{
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "type": "node",
-      "request": "launch",
-      "name": "run file",
-      "runtimeExecutable": "npx",
-      "runtimeArgs": ["c8", "--reporter", "lcov", "node"],
-      "program": "${file}",
-      "args": [],
-      "env": {
-        "NODE_ENV": "development"
-      },
-      "cwd": "${workspaceFolder}",
-      "sourceMaps": true,
-      "outFiles": ["${workspaceFolder}/../*/out/**/*.js"],
-      "preLaunchTask": "build",
-      "outputCapture": "std"
-    }
-  ]
-}
