--- packages/ts/jns42-optimizer/.vscode/tasks.json
@@ -1,15 +1,0 @@
-{
-  "version": "2.0.0",
-  "tasks": [
-    {
-      "type": "npm",
-      "label": "build",
-      "script": "build",
-      "problemMatcher": ["$tsc"],
-      "group": {
-        "kind": "build",
-        "isDefault": true
-      }
-    }
-  ]
-}
