--- packages/ts/jns42-optimizer/src/main.ts
@@ -1,3 +1,0 @@
-export * as schemaTransforms from "./schema-transforms/index.js";
-
-export * from "./schema/index.js";
