--- packages/ts/jns42-optimizer/src/utils/index.ts
@@ -1,10 +1,0 @@
-export * from "./arena.js";
-export * from "./calculate-hash.js";
-export * from "./choose.js";
-export * from "./deep-equal.js";
-export * from "./deep-unique.js";
-export * from "./empty.js";
-export * from "./math.js";
-export * from "./merge.js";
-export * from "./normalize-object.js";
-export * from "./properties.js";
