--- packages/ts/jns42-optimizer/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "@tsconfig/node20",
-  "compilerOptions": {
-    "outDir": "./out",
-    "rootDir": "./src",
-    "sourceMap": true,
-    "declaration": true,
-    "composite": true,
-    "lib": ["ES2023", "DOM"],
-  },
-  "include": ["src"],
-  "references": [{ "path": "../schema-intermediate" }],
-}
